+++ Keysharp.Core/Common/Keyboard/HotstringManager.cs
@@ -11,9 +11,9 @@
 		internal bool hsDoBackspace = true;
 		internal bool hsDoReset;
 		internal bool hsEndCharRequired = true;
-		internal int hsKeyDelay;
+		internal long hsKeyDelay;
 		internal bool hsOmitEndChar;
-		internal int hsPriority;
+		internal long hsPriority;
 		internal bool hsResetUponMouseClick = true;
 		internal bool hsSameLineAction;
 		internal SendModes hsSendMode = SendModes.Input;
